pyproject.toml
[project]
name = "builderstudio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=6.29.5",
    "loguru>=0.7.3",
    "pre-commit>=4.1.0",
    "typing-extensions>=4.12.2",
    "fastapi>=0.115.8",
    "python-dotenv>=1.0.1",
    "commitizen>=4.1.1",
    "pymongo==4.10.0",
    "regex==2024.9.11",
    "sentence-transformers==3.1.1",
    "uvicorn==0.31.0",
    "pydantic==2.5.3",
    "scikit-learn==1.5.2",
    "pyvi==0.1.1",
    "regex==2024.9.11",
    "torch==2.4.1",
    "pytest"
]

[dependency-groups]
dev = [
    "pre-commit",
    "pytest",
    "pytest-cov",
    "pytest-env",
    "pytest-asyncio",
    "ruff>=0.9.4",
]

[tool.ruff]
line-length = 120
target-version = "py311"

exclude = ["frontend"]

[tool.ruff.lint]
select = [
    "F",        # Pyflakes
    "I",        # isort
    "TCH",      # Type-checking imports
    "UP",       # pyupgrade
    "PLC",      # Pylint import conventions
]
extend-select = ["I", "U"]
ignore = ["UP015", "UP035"]  # Keep some reasonable exceptions

[tool.ruff.lint.per-file-ignores]
"*.ipynb" = ["E402", "F811", "E722"]
"__init__.py" = ["F401"]  # Ignore unused imports in __init__ files

[tool.commitizen]
version = "0.1.0"
update_changelog_on_bump = true

[tool.ruff.lint.isort]
combine-as-imports = true
force-wrap-aliases = true
known-first-party = ["backend"]
force-single-line = false
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.lint.pylint]
max-args = 6
